+++ letsencrypt/acme/schemata/challenge.json
@@ -18,7 +18,7 @@
             "type": "array",
             "minItems": 1,
             "items": {
-                "$ref": "file:letsencrypt/client/schemata/challengeobject.json"
+                "$ref": "file:letsencrypt/acme/schemata/challengeobject.json"
             }
         },
         "combinations": {
